changed the folder structure and updated the dashboard design with responsive
