+++ src/pages/dashboard/Dashboard.js
@@ -1,0 +1,7 @@
+import DashboardLayout from "@/components/dashboard/DashboardLayout";
+
+const Dashboard = () => {
+    return <DashboardLayout />
+}
+
+export default Dashboard
