+++ src/pages/homepage/homepage.js
@@ -1,5 +1,6 @@
 const homepage = () => {
 
-}
+};
 
 export default homepage;
+
